backward compactibility fix

+++ src/main/java/io/turntabl/services/PermissionStorage.java
@@ -4,6 +4,7 @@ package io.turntabl.services;
 import io.turntabl.models.PendingRequest;
 import io.turntabl.models.Request;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataAccessException;
 import org.springframework.jdbc.core.BeanPropertyRowMapper;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
@@ -18,7 +19,7 @@ public class PermissionStorage {
 
     public PermissionStorage(){ }
 
-    public long insert(String identifier, String userEmail, Set<String> arnsRequest){
+    public long insert(String userEmail, Set<String> arnsRequest){
         if ( arnsRequest.size() > 0) {
             String arnsString = String.join(" -,,- ", arnsRequest);
 
@@ -27,7 +28,6 @@ public class PermissionStorage {
             insertValue.put("status", "PENDING");
             insertValue.put("useremail", userEmail);
             insertValue.put("arn", arnsString);
-            insertValue.put("identifier", identifier);
 
             Number number = insert.executeAndReturnKey(insertValue);
             return number.longValue();
@@ -38,20 +38,29 @@ public class PermissionStorage {
 
 
     public void approvedRequest( long requestId){
+        Request requestDetails = getRequestDetails(requestId);
+        if ( requestDetails == null){ return ; }
         jdbcTemplate.update("UPDATE requests SET status = 'APPROVED', approvedtime = ? " +
         " WHERE id = ? ", LocalDateTime.now(), requestId
         );
     }
 
     public String removeRequest( long requestId) {
-        String userEmail = getRequestDetails(requestId).getUserEmail();
+        Request requestDetails = getRequestDetails(requestId);
+        if ( requestDetails == null){ return "INVALID"; }
+        String userEmail = requestDetails.getUserEmail();
         jdbcTemplate.update("DELETE FROM requests WHERE id = ?", requestId);
         return userEmail;
     }
 
     public Request getRequestDetails( long requestId) {
-        return jdbcTemplate.queryForObject("SELECT * FROM requests WHERE id = ? AND status = 'PENDING'", new Object[]{requestId},
-                BeanPropertyRowMapper.newInstance(Request.class));
+        try {
+            return jdbcTemplate.queryForObject("SELECT * FROM requests WHERE id = ?", new Object[]{requestId},
+                    BeanPropertyRowMapper.newInstance(Request.class));
+        } catch (DataAccessException e) {
+            e.printStackTrace();
+            return null;
+        }
     }
 
     public List<Request> approvedPermissions(){
@@ -64,15 +73,9 @@ public class PermissionStorage {
                 BeanPropertyRowMapper.newInstance(Request.class))
                  .stream()
                     .map(q -> {
-                        Set<String> awsArns = Arrays.stream(q.getARN().split(" -,,- "))
-                                .map(s -> {
-                                    String[] inter = s.split("/");
-                                    return inter[inter.length - 1];
-                                })
-                                .collect(Collectors.toSet());
-                        return new PendingRequest( q.getUserEmail(), q.getStatus(), q.getIdentifier(), q.getRequest_time(), awsArns);
+                        Set<String> awsArns = new HashSet<>(Arrays.asList(q.getARN().split(" -,,- ")));
+                        return new PendingRequest(q.getUserEmail(), q.getStatus(), awsArns);
                     })
-                 .collect(Collectors.toList());
+                    .collect(Collectors.toList());
     }
 }
-
